rrr: convert misc/alock to module (rrr.alock)
Merges alock.hpp (771 lines) + alock.cpp (785 lines) into the single
module unit src/rrr/misc/alock.cpp (~1515 lines). Exports the
asynchronous-queued-lock surface (ALock and its callback / state
glue) and pulls in rrr.alarm, rrr.callback_wrapper, rrr.dball,
rrr.basetypes, rrr.threading via imports. The sole textual consumer
(rrr.hpp umbrella) now imports rrr.alock instead of including
alock.hpp.

Despite using `Reactor::create_event<IntEvent>()` in its module
purview — the same shape of cross-module template-method call that
crashed clang 22 on fiber_channel — alock compiles cleanly. The
fiber_channel bug must depend on something narrower (perhaps the
`shared_ptr` return type of `create_sp_event`, or the nearby
`event->wait()` virtual call); investigation continues on the
deferred fiber_channel task.
