+++ src/rrr/CMakeLists.txt
@@ -26,6 +26,7 @@ set(RRR_MODULE_SRC
     ${CMAKE_CURRENT_SOURCE_DIR}/base/threading.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/base/unittest.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/misc/alarm.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/misc/alock.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/misc/any_message.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/misc/cpuinfo.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/misc/dball.cpp
